refactor(livechat): collapse ChatFinished into a single function component
Merge the ChatFinished component/container/index into one function
component that uses useTranslation and useStore (dropping the
withTranslation HOC and the separate container). Behavior is unchanged.
Story now injects the finished-conversation messages through the shared
storeDecorator instead of passing props the component no longer takes.
