--- packages/livechat/src/routes/ChatFinished/index.ts
@@ -1,2 +1,0 @@
-export { default as ChatFinished } from './component';
-export { default } from './container';
